[refactor] the message component has been moved to shared.
