+++ client/src/shared/ui/message/index.ts
@@ -1,0 +1,1 @@
+export { Message } from './ui/message';
