+++ client/src/shared/ui/message/model/message.types.ts
@@ -1,0 +1,4 @@
+export interface IMessageProps {
+  text: string;
+  locales: string;
+}
