+++ client/src/shared/ui/message/ui/message.styles.ts
@@ -1,0 +1,9 @@
+import { Typography } from 'antd';
+import styled from 'styled-components';
+
+export const TextStyled = styled(Typography.Text)`
+  text-align: center;
+  font-size: 1.5rem;
+  display: inline-block;
+  width: 100%;
+`;
